+++ release.txt
@@ -1,1 +1,2 @@
-
+Fixes:
+- mds3 tk sample table, format field values are properly aligned with ssm name, and indicate cnv from ssm_id in density mode
