+++ server/src/mds3.gdc.js
@@ -4,7 +4,8 @@ import { compute_bins } from '#shared/termdb.bins.js'
 import { getBin } from '#shared/terms.js'
 import ky from 'ky'
 import nodeFetch from 'node-fetch'
-import { combineSamplesById, guessSsmid } from './mds3.variant2samples.js'
+import { combineSamplesById } from './mds3.variant2samples.js'
+import { guessSsmid } from '#shared/mds3tk.js'
 import { filter2GDCfilter } from './mds3.gdc.filter.js'
 import { write_tmpfile } from './utils.js'
 import { mayLog } from './helpers.ts'
